+++ src/ed.js
@@ -242,7 +242,7 @@ export default class Ed {
     this._insertBlocks(index, toInsert)
     return ids
   }
-  updatePlaceholder (id, status, progress) {
+  updatePlaceholder (id, metadata) {
     let block = this.getBlock(id)
     if (!block) {
       throw new Error('Can not update this placeholder block')
@@ -251,8 +251,10 @@ export default class Ed {
       throw new Error('Block is not a placeholder block')
     }
     // Mutation
+    const {status, progress, failed} = metadata
     if (status != null) block.metadata.status = status
     if (progress != null) block.metadata.progress = progress
+    if (failed != null) block.metadata.failed = failed
     // Let content widgets know to update
     this.trigger('media.update')
     // Let fold media know to update